6. Airflow/requirements.txt
requests
bs4
apache-airflow[celery]==2.5.1